% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC.R
\docType{data}
\name{BC}
\alias{BC}
\title{BC object}
\usage{
## Create FS object
BCobj <- makeBC(RAobj, pedfile, family=NULL, inferSNPs=FALSE,
                filter=list(MAF=0.05, MISS=0.2, BIN=100, DEPTH=5, PVALUE=0.01, MAXDEPTH=500))

## Functions (Methods) of an BC object
BCobj$addBIsnps(parent = "both", LODthres = 10, nComp = 10)
BCobj$addSNPs(LODthres = 10, nComp = 10)
BCobj$computeMap(chrom = NULL, init_r = 0.01, ep = 0.001, method = NULL, err = T, mapped = T, nThreads = 1, rfthres = 0.1)
BCobj$createLG(parent = "both", LODthres = 10, nComp = 10, reset = FALSE)
BCobj$maskSNP(snps)
BCobj$mergeLG(LG, where = NULL, mergeTo = NULL)
BCobj$orderLG(chrom = NULL, mapfun = "haldane", weight = "LOD2", ndim = 30, spar = NULL)
BCobj$plotChr(parent = "maternal", mat = "rf", filename = NULL, chrS = 2, lmai = 0.25)
BCobj$plotLG(parent  = "both", LG = NULL, mat = "rf", filename = NULL, interactive = FALSE, what = NULL)
BCobj$plotLM(LG = NULL, fun = "haldane", col = "black")
BCobj$plotSyn()
BCobj$print(what = NULL, ...)
BCobj$removeLG(LG, where = NULL)
BCobj$removeSNP(snps, where = NULL)
BCobj$rf_2pt(nClust = 2, err = FALSE)
BCobj$unmaskSNP(snps)
BCobj$writeLM(file, direct = "./", LG = NULL, what = NULL)
}
\description{
Class for storing BC data and associated functions for analysis of backcross family populations.
}
\details{
An BC object is created from the \code{\link{makeBC}} function and contains RA data,
various statistics of the dataset that have been computed, and functions (methods)
for analyzing the data. Information in an BC object are specific to backcross family populations.
}
\section{Methods(Functions)}{

\describe{
\item{\code{\link{$addBIsnps}}}{Add Both-Informative (BI) snps to the maternal and/or
paternal linkage groups.}
\item{\code{\link{$addSNPs}}}{Add MI, PI and/or SI SNPs to the existing linkage groups}
\item{\code{\link{$createLG}}}{Create linkage group(s).}
\item{\code{\link{$computeMap}}}{Compute linkage maps for a given marker order.}
\item{\code{\link{$maskSNP}}}{Mask SNP(s) in the dataset.}
\item{\code{\link{$mergeLG}}}{Merge linkage groups.}
\item{\code{\link{$orderLG}}}{Order the SNPs in the linkage group(s).}
\item{\code{\link{$plotChr}}}{Plot the heatmap of the 2-point recombination fraction 
estimates (or LOD scores) when SNPs are ordered according to the genome assembly.}
\item{\code{\link{$plotLG}}}{Plot the heatmap of the 2-point recombination fraction 
estimates (or LOD scores) when SNPs are ordered according to their linkage groups.}
\item{\code{\link{$plotLM}}}{Plot linkage maps.}
\item{\code{\link{$plotSyn}}}{Produce a Synteny plot.}
\item{\code{\link{$print}}}{Print summary information for BC object.}
\item{\code{\link{$removeLG}}}{Remove linkage group(s).}
\item{\code{\link{$removeSNP}}}{Remove SNP(s) from linkage group(s).}
\item{\code{\link{$rf_2pt}}}{Compute the 2-point recombination fraction (and LOD score) between all SNP pairs.}
\item{\code{\link{$unmaskSNP}}}{Unmask SNP(s) in the data set.}
\item{\code{\link{$writeLM}}}{Write linkage mapping results to a file.}
}
}

\author{
Timothy P. Bilton
}
\keyword{datasets}
