% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBIsnps.R
\name{$addBIsnps}
\alias{$addBIsnps}
\title{Add BI SNPs to existing Linkage groups}
\usage{
FSobj$addBIsnps(LODthres = 10, nComp = 30)
}
\arguments{
\item{LODthres}{An positive numeric value for the LOD threshold used in the}

\item{nComp}{An positive integer value}
}
\description{
Method for adding Both-informative (BI) SNPs to paternal and maternal linkage groups
and to combine maternal and paternal linkage groups.
}
\details{
details here
}
\author{
Timothy P. Bilton
}
