% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskSNP.R
\name{$maskSNP}
\alias{$maskSNP}
\alias{$unmaskSNP}
\title{Mask and unmask SNPs}
\usage{
FS$obj$maskSNP(snps)

FS$obj$unmaskSNP(snps)
}
\description{
Method for masking and unmasking SNPs in an FS object.
}
\author{
Timothy P. Bilton
}
