% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLG.R
\name{$mergeLG}
\alias{$mergeLG}
\title{Merge Linkage Groups}
\usage{
FSobj$mergeLG(LG)
}
\arguments{
\item{LG}{An integer vector specifying the indices of the linkage groups to be merged.}
}
\description{
Method for merging multiple linkage group from the list of linkage groups formed
in an FS object.
}
\details{
In a linkage analysis in GUSMap, there may be need to remove linkage groups from the 
list of created linkage groups. The indices of the linkage groups corresponds to the number
given in the output for the FS object. There needs to be at leads two linakge groups 
specified for merging and more than two linkage groups can be merged at once.
}
\author{
Timothy P. Bilton
}
