% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderLG.R
\name{$orderLG}
\alias{$orderLG}
\title{Order linkage groups}
\usage{
FSobj$orderLG(chrom = NULL, mapfun = "haldane", weight="LOD2", ndim=30)
}
\arguments{
\item{chrom}{An integer vector of the indices for the linkage group(s) that are to be ordered.}

\item{mapfun}{A character value for the mapping function to be used}

\item{weight}{A character value for the weight function}

\item{ndim}{A integer value for the number of dimensions to use in the MDS}
}
\description{
Method for ordering linakge groups in GUSMap
}
\details{
Linkage groups created using the \code{\link{$addBIsnps}} are ordered using the multidimensional scaling (MDS)
approach described by \insertCite{preedy2016tag;textual}{GUSMap}.
}
\references{
\insertRef{preedy2016tag}{GUSMap}
}
