% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLG.R
\name{$plotLG}
\alias{$plotLG}
\alias{plotLG}
\title{FS method: Plot linkage groups}
\usage{
plotLG(mat, LG, filename = NULL, names = NULL, chrS = 2, lmai = 2,
  chrom = T, type = "rf")
}
\description{
Plot linkage groups
}
\examples{
## Simulate some sequencing data
set.seed(6745)
config <- list(list(sample(c(1,2,4), size=10, replace=T)), list(sample(c(1,2,4), size=10, replace=T)))
F1data <- simFS(0.01, config=config, meanDepth=10, 
## Compute 2-point recombination fractions
F1data$rf_2pt()
## create paternal and maternal linkage groups
F1data$createLG()

## Plot the linkage groups
F1data$plotLG()

## Plot the linkage groups: suppress interactive plot 
F1data$plotLG(interactive = FALSE)
}
\seealso{
\code{\link{FS}}
}
\author{
Timothy P. Bilton
}
