% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLM.R
\name{$plotLM}
\alias{$plotLM}
\title{FS method: Plot linkage maps}
\usage{
FSobj$plotLM(LG = NULL, fun="haldane", col="black")
}
\arguments{
\item{LG}{An integer vector specifying the indices of the linkage groups to be merged.}

\item{fun}{Character value giving the mapping function to use.}

\item{col}{Vector of character values giving the colors for the linkage maps to produce.}
}
\description{
Method for plotting linkage map in GUSMap
}
\details{
Function produces linkage maps of the combined linkage groups produced from the \code{\link{$addBIsnps}}
add the \code{\link{$orderLG}} functions.
}
\seealso{
\code{\link{FS}}
## Simulate some sequencing data
set.seed(6745)
config <- list(replicate(2,sample(c(1,2,4), size=30, replace=T), simplify=FALSE))
F1data <- simFS(0.01, config=config, meanDepth=10, nInd=50)
## Compute 2-point recombination fractions
F1data$rf_2pt()
## create paternal and maternal linkage groups
F1data$createLG()
## Add BI SNPs
F1data$addBIsnps()
## order Markers
F1data$orderLG(ndim=5)
## Compute Map
F1data$computeMap()

## Produce linkage maps
F1data$plotLM()
}
\author{
Timothy P. Bilton
}
