% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSyn.R
\name{$plotSyn}
\alias{$plotSyn}
\title{BC, FS and IC method: Produce synteny plot}
\usage{
BCobj$plotSyn()
FSobj$plotSyn()
ICobj$plotSyn()
}
\description{
Method for plotting the combined linkage group SNP ordering against the original SNP order
from the genomic assembly.
}
\details{
This function produces a scatter plot of the SNP order as obtained from the linkage analysis
against the SNP order given by the genomic assembly. 

\code{BC} object: The SNP order from the linkage analysis is from the pseudo-testcross linkage groups with BI SNPs
constructed from the \code{\link{addBIsnps}} function and ordered via the \code{\link{$orderLG}} function.

\code{FS} object: The SNP order from the linkage analysis is from the combined linkage groups 
obtained from the \code{\link{$addBIsnps}} function and ordered via the \code{\link{$orderLG}} function.

\code{IC} object: The SNP order from the linkage analysis is from the linkage groups 
computed from the \code{\link{$computeLG}} function and ordered via the \code{\link{$orderLG}} function.
}
\examples{
## Simulate some sequencing data
set.seed(6745)
config <- list(replicate(2,sample(c(1,2,4), size=30, replace=TRUE), simplify=FALSE))
F1data <- simFS(0.01, config=config, meanDepth=10, nInd=50)
## Compute 2-point recombination fractions
F1data$rf_2pt()
## create paternal and maternal linkage groups
F1data$createLG()
## Add BI SNPs
F1data$addBIsnps()
## order Markers
F1data$orderLG(ndim=5)

## Plot synteny plot
F1data$plotSyn()
}
\seealso{
\code{\link{BC}}, \code{\link{FS}}, \code{\link{IC}}
}
\author{
Timothy P. Bilton
}
