% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{$print}
\alias{$print}
\title{FS method: Print summary information}
\usage{
FSobj$print(what = NULL, ...)
}
\arguments{
\item{what}{Character vector giving the type of information to be printed. IF \code{NULL}, then all the 
information is printed.}
}
\description{
Method for printing summary information of an FS object
}
\details{
There are four types of information availabe to be printed for an object (as specified by the \code{what} argument).
\describe{
\item{\code{"data"}}{Summary information of the data as process by the \code{\link{makeFS}} function.}
\item{\code{"LG-pts"}}{Summary information of the pesudo-testcross linkage groups produced by the \code{\link{$createLG}} function.}
\item{\code{"LG-comb"}}{Summary information of the combined linkage groups produced by the \code{\link{$addBIsnps}} function.}
\item{\code{"map"}}{Summary information of the linkage maps produced by the \code{\link{$computeMap}} function.}
}
}
\examples{
## simulate some sequencing data
set.seed(6745)
config <- list(replicate(2,sample(c(1,2,4), size=30, replace=TRUE), simplify=FALSE))
F1data <- simFS(0.01, config=config, meanDepth=10, nInd=50)
# print summary information of data
F1data$print(what = "data")

## Compute 2-point recombination fractions
F1data$rf_2pt()
## create paternal and maternal linkage groups
F1data$createLG()
## Add BI SNPs
F1data$addBIsnps()

# print information of the combined linkage groups
F1data$print(what = "LG-comb")
#print both the data infromation and separate linkage groups
F1data$print(what = c("data", "LG-pts"))
}
\seealso{
\code{\link{FS}}
}
\author{
Timothy P. Bilton
}
