% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLG.R
\name{$removeLG}
\alias{$removeLG}
\title{Remove Linkage Groups}
\usage{
FSobj$removeLG(LG)
}
\arguments{
\item{LG}{An integer vector specifying the number of the linkage groups to be removed}
}
\description{
Method for removing linkage group from the list of linkage groups formed
in an FS object.
}
\details{
In a linkage analysis in GUSMap, there may be need to remove linkage groups from the 
list of created linkage groups. The indices of the linkage groups corresponds to the number
given in the output for the FS object.
}
\author{
Timothy P. Bilton
}
