% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSNP.R
\name{$removeSNP}
\alias{$removeSNP}
\title{Remove SNP from Linkage Group}
\usage{
FSobj$removeSNP(snps)
}
\arguments{
\item{snps}{An integer vector giving the index number of the SNP in the dataset.}
}
\description{
Method for removing removes from creating linkage groups in an FS object.
}
\author{
Timothy P. Bilton
}
