% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_2pt.R
\name{$rf_2pt}
\alias{$rf_2pt}
\alias{comb}
\title{2-point recombination fraction estimates}
\usage{
FSobj$rf_2pt(nClust=2)
}
\description{
Compute 2-point recombination fraction estimates using the method of \insertCite{bilton2018genetics1;textual}{GUSMap}
}
\author{
Timothy P. Bilton
}
