% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSMap-package.R
\docType{package}
\name{GUSMap-package}
\alias{GUSMap-package}
\alias{GUSMap}
\title{Genotyping Uncertainty with Sequencing data for linkage MAPping(GUSMap)}
\description{
Package for (to complete)
}
\details{
\tabular{ll}{ Package: \tab GUSMap\cr Type: \tab Package\cr Version: \tab
1.0.0\cr Date: \tab 2018-08-03\cr License: \tab GPL 3\cr }
}
\references{
\insertAllCited{}
}
\author{
Timothy P. Bilton, Maintainer: Timothy P. Bilton
<tbilton@maths.otago.ac.nz>
}
\keyword{package}
