\name{GusMap-package}
\alias{GusMap,GusMap-package}
\docType{package}
\title{Genotyping with uncertainty and sequening data for linkage mapping}
\description{
Package for performing linkage mapping using outcrossed full-sib family populations that have been sequenced using next generation multiplexing sequencing methods (see Bilton et al. (2017)).
}
\details{
\tabular{ll}{
Package: \tab GusMap\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-07-31\cr
License: \tab GPL 3\cr
}
}
\author{
Timothy P. Bilton,
Maintainer: Timothy P. Bilton <tbilton@maths.otago.ac.nz>
}
\references{
}
\keyword{package}