\name{infer_OPGP_FS}
\alias{infer_OPGP_FS}
\title{To come}
\description{
  To come
}
\usage{
  infer_OPGP_FS <- function(genon, depth, config, ...)
}
\arguments{
  \item{genon}{matrix of genotypes. 2=major homozygous, 1=heterozygous, 0=minor homozygous}
  \item{depth}{matrix of read detphs}
  \item{config}{segregation type for each loci. 1=informative, 2=paternally segregating, 3=maternally segregating, 4=uniformative}
  \item{...}{additional arguments passed  to optim}
}
\details{
}
\value{}
\author{Timothy P. Bilton}
\examples{}
  
  