\name{logit2}
\alias{logit2}
\title{Variation of the logit transformation}
\description{
Functions for transforming and the inverse transforming probabilities according to a variation of the logit transformation
}
\usage{
  logit2(p)
  inv.logit(logit_p)
}
\arguments{
  \item{p}{a numeric vector of probabilities}
  \item{logit_p}{a numeric vector containing the inverse probability values}
}
\details{
  transforms a using the transformation 
}
\value{}
\author{Timothy P. Bilton}
\examples{}

