% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFS.R
\name{makeFS}
\alias{makeFS}
\title{Make a full-sib family (FS) population}
\usage{
makeFS(RAobj, pedfile, family = NULL, filter = list(MAF = 0.05, MISS =
  0.2, BIN = 0, DEPTH = 5, PVALUE = 0.01))
}
\arguments{
\item{RAobj}{Object of class RA created via the \code{\link[GUSbase]{readRA}} function.}

\item{pedfile}{Character string giving the file name (relative to the current directory) of the pedigree file.}

\item{family}{Vector of character strings giving the families to retain in the FS object. This allows a pedigree file with more than one family to be supplied.}

\item{filter}{Named list of thresholds for various filtering criteria.
See below for details.}
}
\value{
An R6 object of class RA
}
\description{
Create an FS object from an RA object, perform standard filtering and computes statistics specific to full-sib family populations.
}
\details{
The segregation type of each SNP is infer based on the genotypes of the parents. The parental genotypes are called homozygous for the 
reference is there is only reference reads seen, heterozygous if at least one read for the reference and alternate allele are seen
and homozygous for the alternate if only reads for the alternate allele is seen. The \code{BIN} filter is implemented to remove any SNPs
with incorrect segregation type. The segregation test procedure for performing the segregation test is described in the supplementary methods 
of the publication by \insertCite{bilton2018genetics1;textual}{GUSMap} (Section 4 of File S1).
}
\section{Filtering}{

The filtering criteria currently implemented are:
\itemize{
\item{Proportion of missing data (\code{BIN}): }{SNPs are binned together if the distance between them is less than threshold value of base pairs (default is 0).
One SNP is then randomly selected from each bin and retained for final analysis. This filtering is to ensure that there is only one SNP on each tag.}
\item{Parental read depth (\code{DEPTH}): }{SNPs are discarded if the read depth of either parent is less than the threshold value (default is 6). 
This filter is to remove SNPs where the parental information is insufficient to infer segregation type.}
\item{Minor allele frequency (\code{MAF}): }{SNPs are discarded if their MAF is less than the threshold (default is 0.05)}
\item{Proportion of missing data (\code{MISS}): }{SNPs are discarded if the proportion of individuals with no reads (e.g. missing genotype)
 is greater than the threshold value (default is 0.5).}
\item{Segregation test P-value (\code{PVALUE}): }{SNPs are discarded if the p-value from a segregation test is smaller than the threshold (default is 0.01).
 This removes SNPs with the wrong segregation type.}
}
}

\section{Pedigree File}{

The pedigree file must be a csv file contains the five columns:
\itemize{
\item SampleID: A unique character string of the sample ID. These correspond to those found in the VCF file
\item IndividualID: A character giving the ID number of the individual for which the sample corresponds to.
Note that some samples can be from the same individual. 
\item Mother: The ID of the mother as given in the IndividualID. Note, if the mother is unknown then this should be left blank.
\item Father: The ID of the father as given in the IndividualID. Note, if the father is unknown then this should be left blank.
\item Family: The name of the Family for a group of progeny with the same parents. Note that this is not necessary (it works all the full-sib families) but if
given must be the same for all the progeny.
}
Grandparents can also be supplied but are only used to infer parental genotypes when the associated read depth is below the threshold \code{DEPTH}.
}

\examples{
## extract filename for Manuka dataset in GUSMap package
vcffile <- Manuka11()

## Convert VCF to RA format
rafile <- VCFtoRA(vcffile$vcf)

## read in the RA data
mkdata <- readRA(rafile)

## Create the FS population
makeFS(mkdata, pedfile=vcffile$ped)
}
\references{
\insertRef{bilton2018genetics1}{GUSMap}
}
\author{
Timothy P. Bilton
}
