% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manuka11.R
\name{Manuka11}
\alias{Manuka11}
\title{Manuka chromosome 11 SNPs and pedigree file}
\usage{
Manuka11()
}
\value{
Function outputs a character string of the complete path to the manuka data set contained in the package
and a pedigree file that goes with the data.
}
\description{
Function for extracting the path to the (VCF) file of the Manuka data used in the publication by Bilton et
al. (2018) and a pedigree file that goes with the data.
}
\details{
The data consists of 680 SNPs, genotyped using genotyping-by-sequencing
methods. The data is in VCF format (see this \href{https://samtools.github.io/hts-specs/VCFv4.2.pdf}{page} 
for specification of VCF format). The pedigree file contains five columns, namely
\itemize{
\item SampleID: A unique character string of the sample ID. These correspond to those found in the VCF file
\item IndividualID: A character giving the ID number of the individual for which the sample corresponds to.
Note that some samples can be from the same individual. 
\item Mother: The ID of the mother as given in the IndividualID. Note, if the mother is unknown then this should be left blank.
\item Father: The ID of the father as given in the IndividualID. Note, if the father is unknown then this should be left blank.
\item Family: The name of the Family for a group of progeny with the same parents. Note that this is not necessary but if
given must be the same for all the progeny.
}
}
\examples{
## extract the name of the vcf fileand the pedigree file
Manuka11()

}
\references{
Bilton, T.P., Schofield, M.R., Black, M.A., Chagné, D., Wilcox,
P.L., & Dodds, K.G. (2018). Accounting for errors in low coverage high-throughput
sequencing data when constructing genetic maps using biparental outcrossed
populations. \emph{Genetics}, \emph{209}(1), 65--76. \url{http://www.genetics.org/content/209/1/65}
}
\author{
Timothy P. Bilton
}
