% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfun.R
\name{mfun}
\alias{mfun}
\title{Mapping functions}
\usage{
mfun(rf, fun = "morgan", centiM = FALSE)
}
\arguments{
\item{rf}{Vector of recombination fraction values}

\item{fun}{Character vector giving the name of the mapping function (see details).}

\item{centiM}{Logical value. If \code{TRUE}, the mapping distances returned are in centiMorgans. Otherwise,
the mapping distances returned are in Morgans.}
}
\description{
Transform recobination fractions into map distances using a specified mapping function.
}
\details{
Currrently, the mapping implemented are: 
\describe{
\item{Morgan mapping function (\code{fun="morgan"})}{\eqn{d = r}}
\item{Haldane mapping function (\code{fun="haldane"})}{\eqn{d = -(1/2)ln(1 - 2r)}}
\item{Kosambi mapping function (\code{fun="kosambi"})}{\eqn{d = (1/4)ln((1+2r)/(1-2r))}}
}
where \eqn{r} denotes the recombination fraction value and \eqn{d} denotes the mapping distance value.
}
\seealso{
\code{\link{$orderLG}}, \code{\link{$plotLM}}
}
\author{
Timothy P. Bilton
}
\keyword{internal}
