% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRA.R
\name{readRA}
\alias{readRA}
\title{Read an Reference/Alternate (RA) file.}
\usage{
readRA(RAfile, pedfile, gform = "reference", sampthres = 0.01,
  filter = list(MAF = 0.05, MISS = 0.2, BIN = 0, DEPTH = 5, PVALUE = 0.01),
  excsamp = NULL)
}
\arguments{
\item{RAfile}{Character string giving the path to the RA file to be read into R. Typically the required string is
returned from the VCFtoRA function when the VCF file is converted to RA format.}

\item{pedfile}{Character string giving the path to the pedigree file of the samples in the RA file. See Detials for more information on specification of this file.}

\item{gform}{Character string specifying whether the SNPs in the RA data have been called using Uneak (\code{gform="uneak"})
or using an reference based assembly (\code{gform="reference"}).}

\item{sampthres}{A numeric value giving the filtering threshold for which infividual samples are removed.}

\item{filter}{A list containing name elements corresponding to the filtering threshold to use for the processing of the full sib families.
See Details for more iinformation regarding the filtering criteria available.}

\item{excsamp}{A character vector of the sample IDs that are to be excluded (or discarded). Note that the sample IDs must correspond
to those given in the RA file that is to be processed.}
}
\value{
A list containing the following elements will be returned;
\itemize{
\item genon: Matrix of genotypes for the simulated sequencing data.
\item depth_Ref: Matrix of the allele counts for the reference allele.
\item depth_Alt: Matrix of the allele counts for the alternate allele.
\item chrom: Vector of the chromosome number corresponding to each SNP as given in the RA file.
\item pos: Vector of chromosome positions corresponding to each SNP as given in the RA file.
\item config: Vector of segregation types used in the simulation.
\item famInfo: A list containing the information of the pedigree structure as supplied in the pedfile file.
simulation.
}
}
\description{
Function which processes an RA file containing full-sib families into R and filters the data.
}
\details{
RA format is a tab-delimited with columns, CHROM, POS, SAMPLES
where SAMPLES consists of sampleIDs, which typically consist of a colon-delimited sampleID, flowcellID, lane, seqlibID.
e.g.,
\tabular{llll}{
CHROM \tab  POS  \tab   999220:C4TWKACXX:7:56 \tab  999204:C4TWKACXX:7:56 \cr
1     \tab  415  \tab   5,0                   \tab  0,3                   \cr
1     \tab  443  \tab   1,0                   \tab  4,4                   \cr
1     \tab  448  \tab   0,0                   \tab  0,2
}

Currently, the data is filtered based on the followin criteria:
\itemize{
\item Minor Allele Frequency (MAF): SNPs with a MAF that is below the specified threshold are discarded.
\item Proportion of missing genotypes (MISS): SNPs where the proportion of genotypes (e.g., non zero read depth)
is less than the specified threshold are discarded.
\item Distance between adjacent SNPs (BIN): SNPs which less than a specified distance apart are binned and one
SNP is retained by random select.
\item Read depth of parental genotypes (DEPTH): Parental genotypes which are (collectly) homozygous are discarded 
if the sum of the read depths is not above the specified threshold. SNPs where the segregation type is not inferred 
are discarded.
\item P-value for segregation test (PVALUE): SNPs must pass a segregation test to be retained in the data set 
for a given p-value. Note that the segregation test is adjusted for low depth as given in the supplementary 
methods of Bilton et al. (2017).
}
}
\examples{
MKfile <- Manuka11()
RAfile <- VCFtoRA(MKfile$vcf, makePed=F)
MKdata <- readRA(RAfile, MKfile$ped)
}
\author{
Timothy P. Bilton
}
