\name{simFS}
\alias{simFS}
\title{Simulation of sequencing data for a single full-sib family.}
\description{
Simulate genotypes and sequencing depth for the progeny of a full-sib family and write the data to various linkage mapping software formats.}
\usage{
  simFS <- function(rVec_f, rVec_m=rVec_f, config, nInd, nSnps, meanDepth, thres=NULL, NoDS=1,
                    formats=list(gusmap=F,onemap=F,lepmap=F,joinmap=F,crimap=F), rd_dist="Neg_Binom",
                    filename=NULL, seed1=1, seed2=1)
}
\arguments{
  \item{rVec_f, rVec_m}{Numeric vector of true paternal and maternal recombinaiton fractions (in the interval [0,0.5]). If only a single number is given, then recombination between adjacent loci are equal.}
  \item{config}{Integer vector of the segregation type for each marker. 1=doubly informative, 2=paternal segregating, 3=maternal segregating and 4=uniformative.}
  \item{nInd}{Positive integer value. The number of individuals.}
  \item{nSnps}{Positive integer value. The number of SNPs (or loci).}
  \item{meanDepth}{Positive numeric value. The mean depth of the read depth distribution.}
  \item{thres}{Numerical Value. Threshold value for which genotype calls with a depth less than the threshold are set to missing.}
  \item{NoDS}{number of data sets to be simulated.}
  \item{formats}{A list object where the element corresponding to the software name is TRUE will produce data in that format. Currently handles software for GusMap, CriMap, JoinMap, OneMap and LepMap2.}
  \item{rd_dist}{String. The distribution for which the read depths are simulated from. Currently, only negative binomial ("Neg_Binom") and Poisson ("Pois") are implemented.}
  \item{filename}{String. Common filename for the data sets generated from the simulation.}
  \item{seed1}{Numeric value. Random seed used for the simulation of the parental phase (or OPGP).}
  \item{seed2}{Numeric value. Random seed used for the simulation of the data sets.}
}
\details{
The \code{simFS} function simulates sequencing data for the progeny of a full-sib family according to the simulation parameters assuming no interference. The processed for the simulation is as follows:
\itemize{
\item{Inhertiance vectors are simulated according to the true recombination fractions, assuming no interference.}
\item{The inheritance vectors are converted to genotypes for the specified OPGP vector (which is simulated based on the \code{config} vector)).}
\item{Seqencing genotypes are generated by first simulating realizations from the read depth distribution and sampling the alleles of the true genotype with equal probability and replacement until a sample size corresponding to the read depth is obtained.}
}

The mean read depth for each loci is assumed to be equal. 

If none of the methods in \code{formats} argument is sepecified to be TRUE, then no files are created and the simulated data is returned to an object (see examples below). Otherwise, datasets are created and written to files in the correct format for the various programs specified.
}
\value{
If at least one method in \code{formats} argument is sepecified to be TRUE, a file containing the simulation parameters is written and files containing the simulated data for each of the methods is written. The naming of the files (if \code{filename="sim"}) is as follows:
\itemize{
\item{Simulation parameters: }{sim_info.txt}
\item{GusMap Genotypes: }{sim_genon_SEQ.txt}
\item{GusMap Read depths: }{sim_depth_SEQ.txt}
\item{OneMap format: }{sim_OneMap.txt}
\item{LepMap2 format: }{sim_LepMap.txt}
\item{CriMap format: }{chr1_sim.gen}
\item{JoinMap format: }{sim_JoinMap.loc}
}

If none of the methods in \code{formats} argument is sepecified to be TRUE, then a list containing the following elements will be returned:
\itemize{
\item{genon: }{Matrix of genotypes for the simulated sequening data.}
\item{depth: }{Matrix of read depths associated with the sequencing genotype calls.}
\item{trueGeno: }{Matrix of true genotypes calls simulated.}
\item{rVec_f: }{Vector of adjacent paternal recombination fractions.}
\item{rVec_m: }{Vector of adjacent maternal recombination fractions.}
\item{nInd: }{Number of individuals.}
\item{nSnps: }{Number of SNPs.}
\item{config: }{Vector of segregation types used in the simulation.}
\item{OPGP: }{Vector of OPGPs used in the simulation.}
\item{meanDepth: }{Mean of the distribution used to simulate the read depths.}
\item{rd_dist: }{Distribution from the which the read detphs where simulated.}
}


}
\author{Timothy P. Bilton}
\examples{
## simulate full sib family
config <- c(2,1,1,3,2,3,1,1,3,1,2,1)
set.seed(3)
F1data <- simFS(runif(11,0.01,0.05), config=config, nInd=50, nSnps=12, meanDepth=5)
## to look at the simulated data and simulation parameters
F1data

## simulate full sib family assuming recombination is equal between loci
config <- c(2,1,1,3,2,3,1,1,3,1,2,1)
F1data <- simFS(0.01, config=config, nInd=50, nSnps=12, meanDepth=5)

## To write the simulated data to files (in GusMap format only). 
config <- c(2,1,1,3,2,3,1,1,3,1,2,1)
simFS(0.01, config=config, nInd=50, nSnps=12, meanDepth=5, formats(gusmap=TRUE))
}
