% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLG.R
\name{$createLG}
\alias{$createLG}
\title{BC and FS method: Create linkage groups}
\usage{
BCobj$createLG(parent = "both", LODthres = 10, nComp = 10)
FSobj$createLG(parent = "both", LODthres = 10, nComp = 10)
}
\arguments{
\item{parent}{A character vector specifying whether to create linakge groups using the maternal-informative (MI)
SNPs (\code{"maternal"}), the paternal-informative (PI) SNPs (\code{"paternal"}), or MI and PI SNPs
(\code{"both"}).}

\item{LODthres}{A positive numeric value specifying the LOD threshold used to add SNPs to the linkage groups.}

\item{nComp}{A positive integer value specifying how many SNPs in the linakge group to compute the average LOD score 
with the unmapped SNP.}
}
\description{
Method for creating maternal and paternal linkage groups using the computed LOD scores.
}
\details{
Linkage groups are formed using the following algorithm
\enumerate{
\item The two unmapped SNPs (e.g. not already in a linkage group) with the highest LOD score are grouped together to
 form a linkage group.
\item The average LOD score between each unmapped SNP and a specified number (\code{nComp}) of SNPs in the 
linkage group that have the highest LOD score with the unmapped SNP is computed. The unmapped SNP with the 
largest average LOD score is mapped to the linkage group if the average LOD score is above the LOD threshold
(given by \code{LODthres}).
\item When no more SNPs are added to the linkage group, repeat steps 1-2 to form a new linkage group.
\item Stop no more linkage groups can be formed or all the SNPs have been mapped 
to a linkage group.
}

The performance of the linkage group algorithm depends on the value of \code{LODthres} and \code{nComp}. The user is
advised to experiment with different values and examine the matrix of recombination fraction estimates (using the \code{\link{$plotLG}}
function).

Notes: 
\itemize{
\item The LOD scores used in this function are computed using the {\code{\link{$rf_2pt}}} which needs to be
run beforehand.
\item A MI linkage group is one which only contains MI SNPs and a PI linkage group is one with only PI SNPs. These 
linkage groups are referred to as the "puesdo-testcross linkage groups".
\item Only MI and PI SNPs are used to construct linkage groups. The SI SNPs inferred in the {\code{\link{makeBC}}} or 
{\code{\link{makeFS}}} function when the argument {\code{inferSNPs = TRUE}} are not used in the {\code{\link{$createLG}}} function.
}
}
\examples{
## Simulate some sequencing data
set.seed(6745)
config <- list(list(sample(c(1,2,4), size=30, replace=TRUE)))
F1data <- simFS(0.01, config=config, meanDepth=10, nInd=50)
## Compute 2-point recombination fractions
F1data$rf_2pt(nClust=1)

## create paternal and maternal linkage groups
F1data$createLG()
}
\seealso{
\code{\link{BC}}, \code{\link{FS}}
}
\author{
Timothy P. Bilton
}
