% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskSNP.R
\name{$maskSNP}
\alias{$maskSNP}
\alias{$unmaskSNP}
\title{BC and FS method: Mask and unmask SNPs}
\usage{
BC$obj$maskSNP(snps)
FS$obj$maskSNP(snps)

BC$obj$unmaskSNP(snps)
FS$obj$unmaskSNP(snps)
}
\arguments{
\item{snps}{An integer vector giving the indices of the SNP(s) in the dataset.}
}
\description{
Method for masking and unmasking SNPs in an FS object.
}
\details{
In the linkage mapping process, some SNPs are often probmatic and there 
may be the need exclude these SNPs in forming linkage groups and computing linkage maps. This can be achieved using
the \code{\link{$maskSNP}} function, which masks SNPs. 
Once a SNP is masked, it will be ignored in the functions \code{\link{$createLG}},
\code{\link{$addBIsnps}}, \code{\link{$computeMap}}. Note, however, that if a SNP is
already included in a linkage group, masking a SNP will not remove it (To remove SNPs from
a linkage group, one needs to use the \code{\link{$removeSNP}} function). 

The \code{\link{$unmaskSNP}} function just unmasks any SNPs which have been masked using 
\code{\link{$maskSNP}}.
}
\examples{
## simulate some sequencing data
set.seed(6745)
config <- list(list(sample(c(1,2,4), size=30, replace=TRUE)))
F1data <- simFS(0.01, config=config, meanDepth=10, nInd=50)

## Mask SNPs 1,5,8,11,13
F1data$maskSNP(c(1,5,8,11))

## Unmask SNPs 5 and 8
F1data$unmaskSNP(c(5,8))

}
\seealso{
\code{\link{BC}}, \code{\link{FS}}
}
\author{
Timothy P. Bilton
}
