% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskSNP.R
\name{$maskSNP}
\alias{$maskSNP}
\alias{$unmaskSNP}
\title{Mask and unmask SNPs}
\usage{
FS$obj$maskSNP(snps)

FS$obj$unmaskSNP(snps)
}
\arguments{
\item{snps}{An integer vector giving the index number of the SNP in the dataset.}
}
\description{
Method for masking and unmasking SNPs in an FS object.
}
\details{
In the linkage mapping process, some SNPs may some themselves to be probmatic and there 
may be the need exclude these SNPs in forming linkage groups etc. This can be achieved using
the \code{\link{$maskSNP}} function, which masks SNPs. 
Once a SNP is masked, it will be ignored in the functions \code{\link{$createLG}},
\code{\link{$addBIsnps}}, \code{\link{$rf_est}}. Note, however, that if a SNP is
already included in a linkage group, masking a SNP will not remove it (To remove SNPs from
one needs to use the \code{\link{$removeSNP}} function).

The \code{\link{$unmaskSNP}} function just unmasks any SNPs which have been masked using 
\code{\link{$maskSNP}}.
}
\author{
Timothy P. Bilton
}
