% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLG.R
\name{$mergeLG}
\alias{$mergeLG}
\title{FS method: Merge Linkage Groups}
\usage{
FSobj$mergeLG(LG, mergeTo = NULL)
}
\arguments{
\item{LG}{An integer vector specifying the indices of the linkage groups to be merged.}

\item{mergeTo}{Character value specifying whether the merged linkage groups are to be considered
as MI linkage groups (\code{maternal}) or PI linkage groups (\code{paternal}).}
}
\description{
Method for merging multiple linkage group from the list of linkage groups contained
in an FS object.
}
\details{
For a linkage analysis in GUSMap, there may be need to merge linkage groups from the 
list of created linkage groups. The indices of the linkage groups corresponds to the number
given in the output for the FS object. There needs to be at leads two linakge groups 
specified for merging and more than two linkage groups can be merged at once.

In the case when at least one MI linkage group is to be merged with at least one PI linkage group, 
the second argument (\code{mergeTo}) specifies whether the merged linkage groups are to be considered
as MI linkage groups (\code{maternal}) or PI linkage groups (\code{paternal}). Otherwise,
the user will be prompted to input which parental line to merge the linkage groups to.
}
\examples{
## simulate sequencing data
config <- list(replicate(2,sample(c(1,2,4), size=30, replace=T), simplify = FALSE))
F1data <- simFS(0.01, config=config, meanDepth=10, nInd=50)

## Compute 2-point recombination fractions
F1data$rf_2pt(nClust=1)

## create paternal and maternal linkage groups
F1data$createLG()
## merge linkage groups 1 and 2
F1data$mergeLG(LG = c(1,2))

## create paternal and maternal linkage groups
F1data$createLG()
## merge linkage groups 1 and 3
F1data$mergeLG(LG = c(1,3))
}
\seealso{
\code{\link{FS}}
}
\author{
Timothy P. Bilton
}
