% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderLG.R
\name{$orderLG}
\alias{$orderLG}
\title{FS method: Order linkage groups}
\usage{
FSobj$orderLG(chrom = NULL, mapfun = "haldane", weight="LOD2", ndim=30)
}
\arguments{
\item{chrom}{An integer vector of the indices for the linkage group(s) that are to be ordered.}

\item{mapfun}{A character value for the mapping function to be used}

\item{weight}{A character value for the weight function}

\item{ndim}{A integer value for the number of dimensions to use in the MDS}
}
\description{
Method for ordering marker in a linakge group
}
\details{
Linkage groups created using the \code{\link{$addBIsnps}} are ordered using the multidimensional scaling (MDS)
approach described by \insertCite{preedy2016tag;textual}{GUSMap}.
}
\examples{
## Simulate some sequencing data
set.seed(6745)
config <- list(list(sample(c(1,2,4), size=10, replace=T)), list(sample(c(1,2,4), size=10, replace=T)))
F1data <- simFS(0.01, config=config, meanDepth=10, 
## Compute 2-point recombination fractions
F1data$rf_2pt()
## create paternal and maternal linkage groups
F1data$createLG()
## Add the BI SNPs
F1data$addBIsnps()

## Order Linkage groups
F1data$orderLG(ndim=5)
}
\references{
\insertRef{preedy2016tag}{GUSMap}
}
\seealso{
\code{\link{FS}}
}
