% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLG.R
\name{$plotLG}
\alias{$plotLG}
\alias{plotLG}
\title{FS method: Heatmap of 2-point rf and LOD estimates from linkage group order}
\usage{
FSobj$plotLG(parent  = "maternal", LG=NULL, mat="rf", filename=NULL, interactive=TRUE, what = NULL)
}
\arguments{
\item{mat}{Charater value for the matrix to be plotted. \code{"rf"} plots the matrix of 2-point 
recombination fractions while \code{"LOD"} plots the matrix of 2-point LOD scores.}

\item{LG}{Integer vector giving the indices of the linkage groups to be plotted.}

\item{filename}{Character value giving the name of the file to save the plot to. If \code{NULL}, the plot is displayed 
in the graphics window and not saved to a file.}

\item{parent}{Character value specifying whether the SNPs segreagting in the maternal parent should be 
plotted (\code{"maternal"}), or whether whether the SNPs segreagting in the paternal parent should be 
plotted (\code{"paternal"}), or whether all the SNPs should be plotted (\code{"both"}).}

\item{interative}{Logical value. If \code{TRUE} then an interactive plot is produced, otherwise a standard
base R plot is used.}

\item{what}{Character vector specifying which list of linkage groups to plot. \code{"LG-pts"} is for 
the pseudo-testcross linkage groups and \code{"LG-comb"} is for the combined linkage groups.}
}
\description{
Method for plotting 2-point recombination fraction and LOD score estimates when the SNPs are ordered according to the linkage groups.
}
\details{
This function plots the heatmap of the matrix of 2-point recombination fraction estimates (or 2-point LOD scores)
as computed from the \code{\link{$rf_2pt}} function. 

The interactive plot (\code{interactive = TRUE}) is produced via the \code{\link[plotly]{plot_ly}} function. Note, however,
that the interactive plot only works for a relatively small number of SNPs (about 1000 is maximum).
Nevertheless, one can still plot each linkage group (using the \code{LG} argument) provided there are 
not too many SNPs in the linkage group.

When \code{what = "LG"}, the pseudo-testcross linkage groups created via the \code{\link{$createLG}} function 
will be plotted. On the other hand, if \code{what = "LG_BI"}, then the combined linkage groups created from 
the \code{\link{$addBIsnps}} function will be plotted. When \code{what = NULL}, the combined linkage groups will
be plotted if available, otherwise the pseudo-testcross linkage groups will be plotted.
}
\examples{
## Simulate some sequencing data
set.seed(6745)
config <- list(replicate(2, sample(c(1,2,4), size=30, replace=TRUE), simplify=FALSE))
F1data <- simFS(0.01, config=config, meanDepth=10, nInd=50)
## Compute 2-point recombination fractions
F1data$rf_2pt()
## create paternal and maternal linkage groups
F1data$createLG()

## Plot the linkage groups
F1data$plotLG()

## Plot the linkage groups: suppress interactive plot 
F1data$plotLG(interactive = FALSE)
}
\seealso{
\code{\link{FS}}
}
\author{
Timothy P. Bilton
}
