% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLM.R
\name{$plotLM}
\alias{$plotLM}
\title{BC, FS and IC method: Plot linkage maps}
\usage{
BCobj$plotLM(LG = NULL, fun="haldane", col="black")
FSobj$plotLM(LG = NULL, fun="haldane", col="black")
ICobj$plotLM(LG = NULL, fun="haldane", col="black")
}
\arguments{
\item{LG}{An integer vector specifying the indices of the linkage groups for which their linkage maps are to be plotted.}

\item{fun}{Character value giving the mapping function to use which is passed to \code{\link{mfun}}.}

\item{col}{Vector of character values giving the colors of the linkage maps.}
}
\description{
Method for plotting linkage maps in GUSMap
}
\details{
Function displays linkage maps computed using the \code{\link{$computeMap}} function.

For a list of mapping functions available, see \code{\link{mfun}}.
}
\examples{
## Simulate some sequencing data
set.seed(6745)
config <- list(replicate(2,sample(c(1,2,4), size=30, replace=TRUE), simplify=FALSE))
F1data <- simFS(0.01, config=config, meanDepth=10, nInd=50)
## Compute 2-point recombination fractions
F1data$rf_2pt()
## create paternal and maternal linkage groups
F1data$createLG()
## Add BI SNPs
F1data$addBIsnps()
## order Markers
F1data$orderLG(ndim=5)
## Compute Map
F1data$computeMap()

## Produce linkage maps
F1data$plotLM()
}
\seealso{
\code{\link{BC}}, \code{\link{FS}}, \code{\link{IC}}, \code{\link{mfun}}
}
\author{
Timothy P. Bilton
}
