% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSNP.R
\name{$removeSNP}
\alias{$removeSNP}
\title{FS method: Remove SNP from Linkage Group}
\usage{
FSobj$removeSNP(snps, where = NULL)
}
\arguments{
\item{snps}{An integer vector giving the indices of the SNP to be removed.}

\item{where}{Character vector specifying which set of linkage groups to remove SNPs from (see details).}
}
\description{
Method for removing SNPs from linkage groups in an FS object.
}
\details{
For a linkage analysis in GUSMap, there may be a need to delete linkage groups from the
list of created linkage groups. The indices of the linkage groups corresponds to the number
given in the output for the FS object. More than one linkage group can be deleted at once.

Note: When the linkage group(s) are deleted, the indices of the remaining linkage groups 
changes. Print the FS object to investiage what these after running this function.

When \code{where = "LG"}, SNPs will be removed from the paternal and maternal linkage 
groups created via the \code{\link{$createLG}} function. On the other hand, if 
\code{where = "LG_BI"}, then SNPs will be removed from the combined linkage groups 
created from the \code{\link{$addBIsnps}} function (e.g., the linkage groups which include the BI SNPs).
When \code{where = NULL}, SNPs will be removed from the combined linkage groups if availabve,
otherwise they will be removed from the maternal and paternal linkage groups.
}
\examples{
## Simulate some sequencing data
set.seed(6745)
config <- list(list(sample(c(1,2,4), size=20, replace=T)))
F1data <- simFS(0.01, config=config, meanDepth=10, nInd=50) 
## Compute 2-point recombination fractions
F1data$rf_2pt()
## create paternal and maternal linkage groups
F1data$createLG()

## Remove SNPs 3,7,15,20 from the linkage groups
F1data$removeSNP(snps = c(3,7,15,20))
}
\seealso{
\code{\link{FS}}
}
\author{
Timothy P. Bilton
}
