% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_2pt.R
\name{$rf_2pt}
\alias{$rf_2pt}
\alias{comb}
\title{FS method: Compute 2-point recombination fraction estimates and LOD scores}
\usage{
FSobj$rf_2pt(nClust = 2, err = FALSE)
}
\arguments{
\item{nClust}{An integer value for the number of cores to use in the parallelization of 
computing the 2-point recombination fraction estimates.}

\item{err}{Locial value. If \code{TRUE}, a sequencing error parameter is also estimated as 
part of the 2-point recombination fraction estimates. If \code{FALSE}, then the sequencing 
error parameter is fixed at zero.}
}
\description{
Method for estimating 2-point recombination fraction and associated LOD scores.
}
\details{
In this function, the recombination fraction for all pairs of SNPs is computed. Estimation 
is performed by optimimizing the hidden Markov model (HMM) likelihood given by 
\insertCite{bilton2018genetics1;textual}{GUSMap} for two markers, where the parental phase 
that is used is the one that maximizes the likelihood, given the segregation type that has been 
inferred. The \code{err} specifies whether sequencing errors should
also be estimated when computing the 2-point recombination fraction estimates. LOD scores associated 
with each pair of SNPs are also computed.

Parallelization is used to speed up the estimation process using the \code{\link[foreach]{foreach}}
function. The argument \code{nClust} specifies the number of cores to use in the parallelization. 
Note: It is important that the number of cores you specify is not more than what is available on 
your comupter (otherwise bad things can happen!).

Currently, only 2-point recombination fractions for a single full-sib family can be computed. However,
there are plans to extend this function to multiple full-sib families in the future.
 
Note: This function can take a while, espically when there are a large number of SNPs.
}
\examples{
## simulate some sequencing data
set.seed(6745)
config <- list(list(sample(c(1,2,4), size=30, replace=T)))
F1data <- simFS(0.01, config=config, meanDepth=10, nInd=50)

## Compute 2-point recombination fractions
F1data$rf_2pt(nClust=1)
}
\references{
\insertRef{bilton2018genetics1}{GUSMap}
}
\seealso{
\code{\link{FS}}
}
\author{
Timothy P. Bilton
}
