% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\docType{data}
\name{FS}
\alias{FS}
\title{FS object}
\usage{
## Create FS object
FSobj <- makeFS(RAobj, pedfile, family=NULL, 
                filter=list(MAF=0.05, MISS=0.2, BIN=0, DEPTH=5, PVALUE=0.01))

## Functions (Methods) of an FS object
FSobj$addBIsnps(LODthres = 10, nComp = 10)
FSobj$createLG(parent = "both", LODthres = 10, nComp = 10)
FSobj$maskSNP(snps)
FSobj$mergeLG(LG)
FSobj$orderLG(chrom = NULL, mapfun = "haldane",
                weight="LOD2", ndim=30)
FSobj$plotChr(  )
FSobj$plotLG(  )
FSobj$removeLG(LG)
FSobj$removeSNP(snps)
FSobj$rf_2pt(nClust = 2)
FSobj$unmaskSNP(snps)
FSobj$write()
}
\description{
Class for storing RA data and associated functions for analysis with full-sib populations.
}
\details{
An FS object is created from the \code{\link{makeFS}} function and contains RA data,
various statistics of the dataset that have been computed, and functions (methods)
for analyzing the data. Information in an FS object are specific to full-sib family populations.
}
\section{Methods(Functions)}{

\describe{
\item{\code{\link{$addBIsnps}}}{Add Both-Informative (BI) snps to the maternal and/or
paternal linkage groups and merge the parental linkage groups.}
\item{\code{\link{$createLG}}}{Create linkage group(s).}
\item{\code{\link{$maskSNP}}}{Mask SNP(s) in the dataset.}
\item{\code{\link{$mergeLG}}}{Merge linkage groups.}
\item{\code{\link{$orderLG}}}{Order the SNPs in the linkage group(s).}
\item{\code{\link{$plotChr}}}{Plot the heatmap of the 2-point recombination fraction 
estimates (or LOD scores) when SNPs are ordered according to the genome assembly.}
\item{\code{\link{$plotLG}}}{Plot the heatmap of the 2-point recombination fraction 
estimates (or LOD scores) when SNPs are ordered according to their linkage groups.}
\item{\code{\link{$removeLG}}}{Remove linkage group(s).}
\item{\code{\link{$removeSNP}}}{Remove SNP(s) from linkage group(s).}
\item{\code{\link{$rf_2pt}}}{Compute the 2-point recombination fraction (and LOD score) between all SNP pairs.}
\item{\code{\link{$unmaskSNP}}}{Unmask SNP(s) in the data set.}
}
}

\author{
Timothy P. Bilton
}
\keyword{datasets}
