% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\docType{data}
\name{FS}
\alias{FS}
\title{FS object}
\usage{
## Create FS object
FSobj <- makeFS(RAobj, pedfile, family=NULL, 
                filter=list(MAF=0.05, MISS=0.2, BIN=0, DEPTH=5, PVALUE=0.01))

## Functions (Methods) of FS object
FSobj$maskSNP(snps)
FSobj$rf_2pt(nClust=2)
FSobj$unmaskSNP(snps)
}
\description{
Class for storing RA data and associated functions for analysis with full-sib populations.
}
\details{
An FS object is created from the \code{\link{makeFS}} function and contains RA data,
various statistics of the dataset that have been computed, and functions (methods)
for analyzing the data. Information in an FS object are specific to full-sib family populations.
}
\section{Methods(Functions)}{

\describe{
\item{\code{\link{$maskSNP}}}{Mask SNPs from the data set.}
\item{\code{\link{$rf_2pt}}}{Compute the 2-point recombination fraction between all SNP pairs.}
\item{\code{\link{$unmaskSNP}}}{Unmask SNPs from the data set.}
}
}

\author{
Timothy P. Bilton
}
\keyword{datasets}
