% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{FS}
\alias{FS}
\title{FS object}
\usage{
## Create FS object
FSobj <- makeFS(RAobj, pedfile, family=NULL, inferSNPs = FALSE,
                filter=list(MAF=0.05, MISS=0.2, BIN=100, DEPTH=5, PVALUE=0.01, MAXDEPTH=500))

## Functions (Methods) of an FS object
FSobj$addBIsnps(LODthres = 10, nComp = 10)
FSobj$computeMap(chrom=NULL, init_r=0.01, ep=0.001, method="optim", sexSpec=F, err=T, mapped=T, nThreads=1)
FSobj$createLG(parent = "both", LODthres = 10, nComp = 10)
FSobj$maskSNP(snps)
FSobj$mergeLG(LG, where = NULL, mergeTo = NULL)
FSobj$orderLG(chrom = NULL, mapfun = "haldane", weight="LOD2", ndim=30, spar = NULL)
FSobj$plotChr(parent = "maternal", mat="rf", filename=NULL, chrS=2, lmai=2)
FSobj$plotLG(parent  = "maternal", LG=NULL, mat="rf", filename=NULL, interactive=TRUE, what = NULL)
FSobj$plotLM(LG = NULL, fun="haldane", col="black")
FSobj$plotSyn()
FSobj$print(what = NULL, ...)
FSobj$removeLG(LG, where = NULL)
FSobj$removeSNP(snps, where = NULL)
FSobj$rf_2pt(nClust = 2, err=FALSE)
FSobj$unmaskSNP(snps)
FSobj$writeLM(file, direct = "./", LG = NULL, what = NULL)
}
\description{
Class for storing RA data and associated functions for analysis of full-sib family populations.
}
\details{
An FS object is created from the \code{\link{makeFS}} function and contains RA data,
various statistics of the dataset that have been computed, and functions (methods)
for analyzing the data. Information in an FS object are specific to full-sib family populations.
}
\section{Methods(Functions)}{

\describe{
\item{\code{\link{$addBIsnps}}}{Add Both-Informative (BI) snps to the maternal and/or
paternal linkage groups and merge the parental linkage groups.}
\item{\code{\link{$addSNPs}}}{Add MI, PI and/or SI SNPs to the existing linkage groups}
\item{\code{\link{$createLG}}}{Create linkage group(s).}
\item{\code{\link{$computeMap}}}{Compute linkage maps for a given marker order.}
\item{\code{\link{$maskSNP}}}{Mask SNP(s) in the dataset.}
\item{\code{\link{$mergeLG}}}{Merge linkage groups.}
\item{\code{\link{$orderLG}}}{Order the SNPs in the linkage group(s).}
\item{\code{\link{$plotChr}}}{Plot the heatmap of the 2-point recombination fraction 
estimates (or LOD scores) when SNPs are ordered according to the genome assembly.}
\item{\code{\link{$plotLG}}}{Plot the heatmap of the 2-point recombination fraction 
estimates (or LOD scores) when SNPs are ordered according to their linkage groups.}
\item{\code{\link{$plotLM}}}{Plot linkage maps.}
\item{\code{\link{$plotSyn}}}{Produce a Synteny plot.}
\item{\code{\link{$print}}}{Print summary information for FS object.}
\item{\code{\link{$removeLG}}}{Remove linkage group(s).}
\item{\code{\link{$removeSNP}}}{Remove SNP(s) from linkage group(s).}
\item{\code{\link{$rf_2pt}}}{Compute the 2-point recombination fraction (and LOD score) between all SNP pairs.}
\item{\code{\link{$unmaskSNP}}}{Unmask SNP(s) in the data set.}
\item{\code{\link{$writeLM}}}{Write linkage mapping results to a file.}
}
}

\author{
Timothy P. Bilton
}
\section{Super class}{
\code{\link[GUSbase:RA]{GUSbase::RA}} -> \code{FS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FS$new()}}
\item \href{#method-print}{\code{FS$print()}}
\item \href{#method-removeSNP}{\code{FS$removeSNP()}}
\item \href{#method-removeLG}{\code{FS$removeLG()}}
\item \href{#method-mergeLG}{\code{FS$mergeLG()}}
\item \href{#method-maskSNP}{\code{FS$maskSNP()}}
\item \href{#method-unmaskSNP}{\code{FS$unmaskSNP()}}
\item \href{#method-rf_2pt}{\code{FS$rf_2pt()}}
\item \href{#method-createLG}{\code{FS$createLG()}}
\item \href{#method-addSNPs}{\code{FS$addSNPs()}}
\item \href{#method-addBIsnps}{\code{FS$addBIsnps()}}
\item \href{#method-orderLG}{\code{FS$orderLG()}}
\item \href{#method-plotLG}{\code{FS$plotLG()}}
\item \href{#method-plotChr}{\code{FS$plotChr()}}
\item \href{#method-plotLM}{\code{FS$plotLM()}}
\item \href{#method-plotSyn}{\code{FS$plotSyn()}}
\item \href{#method-computeMap}{\code{FS$computeMap()}}
\item \href{#method-writeLM}{\code{FS$writeLM()}}
\item \href{#method-cometPlot}{\code{FS$cometPlot()}}
\item \href{#method-rocketPlot}{\code{FS$rocketPlot()}}
\item \href{#method-RDDPlot}{\code{FS$RDDPlot()}}
\item \href{#method-clone}{\code{FS$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="extractVar">}\href{../../GUSbase/html/RA.html#method-extractVar}{\code{GUSbase::RA$extractVar()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="mergeSamples">}\href{../../GUSbase/html/RA.html#method-mergeSamples}{\code{GUSbase::RA$mergeSamples()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="writeRA">}\href{../../GUSbase/html/RA.html#method-writeRA}{\code{GUSbase::RA$writeRA()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="writeVCF">}\href{../../GUSbase/html/RA.html#method-writeVCF}{\code{GUSbase::RA$writeVCF()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$new(R6obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$print(what = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removeSNP"></a>}}
\if{latex}{\out{\hypertarget{method-removeSNP}{}}}
\subsection{Method \code{removeSNP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$removeSNP(snps, where = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removeLG"></a>}}
\if{latex}{\out{\hypertarget{method-removeLG}{}}}
\subsection{Method \code{removeLG()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$removeLG(LG, where = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mergeLG"></a>}}
\if{latex}{\out{\hypertarget{method-mergeLG}{}}}
\subsection{Method \code{mergeLG()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$mergeLG(LG, where = NULL, mergeTo = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-maskSNP"></a>}}
\if{latex}{\out{\hypertarget{method-maskSNP}{}}}
\subsection{Method \code{maskSNP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$maskSNP(snps)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unmaskSNP"></a>}}
\if{latex}{\out{\hypertarget{method-unmaskSNP}{}}}
\subsection{Method \code{unmaskSNP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$unmaskSNP(snps)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rf_2pt"></a>}}
\if{latex}{\out{\hypertarget{method-rf_2pt}{}}}
\subsection{Method \code{rf_2pt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$rf_2pt(nClust = 2, err = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createLG"></a>}}
\if{latex}{\out{\hypertarget{method-createLG}{}}}
\subsection{Method \code{createLG()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$createLG(parent = "both", LODthres = 10, nComp = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addSNPs"></a>}}
\if{latex}{\out{\hypertarget{method-addSNPs}{}}}
\subsection{Method \code{addSNPs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$addSNPs(LODthres = 10, nComp = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addBIsnps"></a>}}
\if{latex}{\out{\hypertarget{method-addBIsnps}{}}}
\subsection{Method \code{addBIsnps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$addBIsnps(LODthres = 10, nComp = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderLG"></a>}}
\if{latex}{\out{\hypertarget{method-orderLG}{}}}
\subsection{Method \code{orderLG()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$orderLG(
  LG = NULL,
  mapfun = "haldane",
  weight = "LOD2",
  ndim = 30,
  spar = NULL,
  filename = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotLG"></a>}}
\if{latex}{\out{\hypertarget{method-plotLG}{}}}
\subsection{Method \code{plotLG()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$plotLG(
  parent = "maternal",
  LG = NULL,
  mat = "rf",
  filename = NULL,
  interactive = TRUE,
  what = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotChr"></a>}}
\if{latex}{\out{\hypertarget{method-plotChr}{}}}
\subsection{Method \code{plotChr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$plotChr(
  chrom = NULL,
  parent = "maternal",
  mat = c("rf"),
  filename = NULL,
  chrS = 2,
  lmai = 2
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotLM"></a>}}
\if{latex}{\out{\hypertarget{method-plotLM}{}}}
\subsection{Method \code{plotLM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$plotLM(LG = NULL, fun = "haldane", col = "black")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSyn"></a>}}
\if{latex}{\out{\hypertarget{method-plotSyn}{}}}
\subsection{Method \code{plotSyn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$plotSyn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-computeMap"></a>}}
\if{latex}{\out{\hypertarget{method-computeMap}{}}}
\subsection{Method \code{computeMap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$computeMap(
  chrom = NULL,
  init_r = 0.001,
  ep = 0.001,
  method = "optim",
  sexSpec = FALSE,
  err = TRUE,
  multiErr = FALSE,
  mapped = TRUE,
  nThreads = 1,
  inferOPGP = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-writeLM"></a>}}
\if{latex}{\out{\hypertarget{method-writeLM}{}}}
\subsection{Method \code{writeLM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$writeLM(file, direct = "./", LG = NULL, what = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cometPlot"></a>}}
\if{latex}{\out{\hypertarget{method-cometPlot}{}}}
\subsection{Method \code{cometPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$cometPlot(
  filename = NULL,
  cex = 1,
  maxdepth = 500,
  maxSNPs = 1e+05,
  res = 300,
  color = NULL,
  ind = FALSE,
  ncores = 1,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rocketPlot"></a>}}
\if{latex}{\out{\hypertarget{method-rocketPlot}{}}}
\subsection{Method \code{rocketPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$rocketPlot(
  ploid = 2,
  filename = NULL,
  cex = 1,
  maxdepth = 500,
  maxSNPs = 1e+05,
  res = 300,
  scaled = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RDDPlot"></a>}}
\if{latex}{\out{\hypertarget{method-RDDPlot}{}}}
\subsection{Method \code{RDDPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$RDDPlot(filename = NULL, maxdepth = 500, maxSNPs = 1e+05, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
