% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSMap-package.R
\docType{package}
\name{GUSMap-package}
\alias{GUSMap-package}
\alias{GUSMap}
\title{Genotyping Uncertainty with Sequencing data for linkage Mapping (GUSMap)}
\description{
Package for constructing genetic maps using outcrossed full-sib family
populations that have been sequenced using next generation multiplexing
sequencing methods (see Bilton et al. (2017)).
}
\details{
\tabular{ll}{ Package: \tab GUSMap\cr Type: \tab Package\cr Version: \tab
0.1.1\cr Date: \tab 2018-06-02\cr License: \tab GPL 3\cr }
}
\references{
Bilton, T.P., Schofield, M.R., Black, M.A., Chagné, D., Wilcox,
P.L., Dodds K.G. (2017). Accounting for errors in low coverage high-throughput
sequencing data when constructing genetic maps using biparental outcrossed
populations. Unpublished Manuscript.
}
\author{
Timothy P. Bilton, Maintainer: Timothy P. Bilton
<tbilton@maths.otago.ac.nz>
}
\keyword{package}
