% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSMap-package.R
\docType{package}
\name{GUSMap-package}
\alias{GUSMap-package}
\alias{GUSMap}
\title{Genotyping Uncertainty with Sequencing data for linkage Mapping (GUSMap)}
\description{
Package for constructing genetic maps using outcrossed full-sib family
populations that have been sequenced using next generation multiplexing
sequencing methods \insertCite{@see @bilton2018genetics1}{GUSMap}.
}
\details{
\tabular{ll}{ Package: \tab GUSMap\cr Type: \tab Package\cr Version: \tab
1.1.0\cr Date: \tab 2019-03-12\cr License: \tab GPL 3\cr }
}
\references{
\insertAllCited{}
}
\author{
Timothy P. Bilton, Maintainer: Timothy P. Bilton
<tbilton@maths.otago.ac.nz>
}
\keyword{package}
