\name{GusMap-package}
\alias{GusMap}
\docType{package}
\title{Genotyping Uncertainty with Sequening data for linkage MAPping (GusMap)}
\description{
Package for constructing genetic maps using outcrossed full-sib family populations that have been sequenced using next generation multiplexing sequencing methods (see Bilton et al. (2017)).
}
\details{
\tabular{ll}{
Package: \tab GusMap\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-08-31\cr
License: \tab GPL 3\cr
}
}
\author{
Timothy P. Bilton,
Maintainer: Timothy P. Bilton <tbilton@maths.otago.ac.nz>
}
\references{
Bilton, T.P., Schofield, M.R., Black, M.A., Chagne D., Wilcox P., Dodds K.G. (2017). More accurate genetic maps using low coverage sequencing data in multiparental outcrossed populations. Unpublished manuscript.

Dodds K.G., McEwan J.C., Brauning R., Anderson R.M., van Stijn, T.C., Kristjansson, T., Clarke, S.M. (2015). Construction of relatedness matrices using genotyping-by-sequencing data. \emph{BMC Genomics}. 16:1047.
}
\keyword{package}