\name{GUSMap-package}
\alias{GUSMap}
\docType{package}
\title{Genotyping Uncertainty with Sequening data for linkage MAPping (GUSMap)}
\description{
Package for constructing genetic maps using outcrossed full-sib family populations that have been sequenced using next generation multiplexing sequencing methods (see Bilton et al. (2017)).
}
\details{
\tabular{ll}{
Package: \tab GUSMap\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2017-10-31\cr
License: \tab GPL 3\cr
}
}
\author{
Timothy P. Bilton,
Maintainer: Timothy P. Bilton <tbilton@maths.otago.ac.nz>
}
\references{
Bilton, T.P., Schofield, M.R., Black, M.A., Chagne D., Wilcox P., Dodds K.G. (2017). Accounting for errors in low coverage sequencing data when constructing genetic maps using multiparental outcrossed populations. Unpublished Manuscript
}
\keyword{package}