% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manuka11.R
\name{Manuka11}
\alias{Manuka11}
\title{Manuka chromosome 11 SNPs}
\usage{
Manuka11()
}
\value{
Function outputs outputs a character sting of the complete path to the data set in the package.
}
\description{
Function for extracting the path to the (VCF) file of the Manuka data used in the publication by Bilton et
al. (2017).
}
\details{
The data consists of 680 SNPs, genotyped using genotyping-by-sequencing
methods. The data is in VCF format (see this \href{https://samtools.github.io/hts-specs/VCFv4.2.pdf}{page} 
for specification of VCF format).
}
\examples{

## extract the name of the vcf file 
vcffile <- Manuka()

}
\references{
Bilton, T.P., Schofield, M.R., Black, M.A., Chagné;, D., Wilcox,
P.L., Dodds K.G. (2017). Accounting for errors in low coverage high-throughput
sequencing data when constructing genetic maps using biparental outcrossed
populations. Unpublished Manuscript.
}
\author{
Timothy P. Bilton
}
