% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRA.R
\name{readRA}
\alias{readRA}
\title{Read an Reference/Alternate (RA) file.}
\usage{
readRA(genofile, gform, sampthres = 0.01, excsamp = NULL)
}
\arguments{
\item{gform}{Character string specifying whether the SNPs in the RA data have been called using Uneak (\code{gform="uneak"})
or using an reference based assembly (\code{gform="reference"}).}

\item{sampthres}{A numeric value giving the filtering threshold for which individual samples are removed.}

\item{excsamp}{A character vector of the sample IDs that are to be excluded (or discarded). Note that the sample IDs must correspond
to those given in the RA file that is to be processed.}

\item{RAfile}{Character string giving the path to the RA file to be read into R. Typically the required string is
returned from the VCFtoRA function when the VCF file is converted to RA format.}
}
\value{
An R6 object of class RA.
}
\description{
Function which processes an RA file into R.
}
\details{
RA format is a tab-delimited with columns, CHROM, POS, SAMPLES
where SAMPLES consists of sampleIDs, which typically consist of a colon-delimited sampleID, flowcellID, lane, seqlibID.
e.g.,
\tabular{llll}{
CHROM \tab  POS  \tab   999220:C4TWKACXX:7:56 \tab  999204:C4TWKACXX:7:56 \cr
1     \tab  415  \tab   5,0                   \tab  0,3                   \cr
1     \tab  443  \tab   1,0                   \tab  4,4                   \cr
1     \tab  448  \tab   0,0                   \tab  0,2
}
}
\examples{
MKfile <- Manuka11()
RAfile <- VCFtoRA(MKfile$vcf, makePed=F)
MKdata <- readRA(RAfile, MKfile$ped)
}
\author{
Timothy P. Bilton
}
