% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRA.R
\name{readRA}
\alias{readRA}
\title{Read in an Reference/Alternate (RA) file.}
\usage{
readRA(RAfile, gform = "reference", pedfile, sampthres = 0.01,
  filter = list(MAF = 0.05, MISS = 0.2, BIN = 0, DEPTH = 5, PVALUE = 0.01),
  excsamp = NULL)
}
\arguments{
\item{RAfile}{Character string giving the path to the RA file to be read into R. Typically the required string is
returned from the VCFtoRA function when the VCF file is converted to RA format.}

\item{gform}{Character string specifying whether the SNPs in the RA data have been called de novo (\code{gform="denovo"})
or using an reference based assembly (\code{gform="reference"}).}

\item{pedfile}{Character string giving the path to the pedigree file of the samples in the RA file. See Detials for more information on specification of this file.}

\item{sampthres}{filtering to finish}
}
\description{
Function which reads in an RA file into R.
}
\author{
Timothy P. Bilton
}
