% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFtoRA.R
\name{VCFtoRA}
\alias{VCFtoRA}
\title{Convert VCF file into RA (Reference/Alternative) file.}
\usage{
VCFtoRA(infilename, direct = "./")
}
\arguments{
\item{infilename}{String giving the filename of the VCF file to be converted to RA format}

\item{direct}{String of the directory (or relative to the working direct) where the RA file is to be written.}
}
\value{
A string of the complete file path and name of the RA file created from the function.
In addition to creating a RA file, a pedigree file is also initialized in the same folder as the RA file.
}
\description{
Function for converting a VCF file into RA format.
}
\details{
RA format is a tab-delimited with columns, CHROM, POS, SAMPLES
where SAMPLES consists of sampleIDs, which typically consist of a colon-delimited sampleID, flowcellID, lane, seqlibID.
e.g.,
\tabular{llll}{
CHROM \tab  POS  \tab   999220:C4TWKACXX:7:56 \tab  999204:C4TWKACXX:7:56 \cr
1     \tab  415  \tab   5,0                   \tab  0,3                   \cr
1     \tab  443  \tab   1,0                   \tab  4,4                   \cr
1     \tab  448  \tab   0,0                   \tab  0,2
}
Note: Indels are removed, multiple alternative alleles are removed and ./. is translated into 0,0.
}
\author{
Timothy P. Bilton. Adapted from a Python script written by Rudiger Brauning and Rachael Ashby.
}
